Revert "Implement @std/esm ES6 import/exporty code"
This reverts commit 9fdaed09cc54a35b7f4c7ddcedce41d206833b04.
