+++ .gitignore
@@ -1,4 +1,3 @@
-.esm-cache
 package-lock.json
 env
 .nyc_output
