+++ packages/turf-invariant/package.json
@@ -3,11 +3,9 @@
   "version": "4.7.3",
   "description": "turf invariant module",
   "main": "index.js",
-  "module": "index.mjs",
   "types": "index.d.ts",
   "files": [
     "index.js",
-    "index.mjs",
     "index.d.ts"
   ],
   "scripts": {
@@ -33,8 +31,5 @@
     "@turf/helpers": "^4.7.3",
     "benchmark": "2.1.4",
     "tape": "4.8.0"
-  },
-  "dependencies": {
-    "@std/esm": "^0.10.2"
   }
 }
